Finished initial top down mode.
A lot was added here. Working with unreal changed the way I think about components, so I added a direction component to wrap the NDC, and an AimToMouseComponent that automatically aims the direction towards the mouse. Added a ParentComponent which links an entity in space to another entity (may need to be renamed)