Patched NDC to not copy the datamap when they are move constructed
This also meant patching a bug that allowed the move construction of elements when Pools are resized.
Planning to add a hard "entity limit" to the engine. This would allow further optimization by allowing all components to be alloced up front, and allowing ptrs to comps so they don't ahve to be retrieved manually.